created all apis, testes basic apis

+++ src/main/java/com/aayush/food_order_app/model/Cart.java
@@ -21,7 +21,7 @@ public class Cart
     @OneToOne
     private User customer;
 
-    private long total;
+    private double total;
 
     @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<CartItem> items = new ArrayList<>();
